requirements.txt
Django==3.2.6
django-elasticsearch-dsl==7.2.0
djangorestframework==3.12.4
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
